requirements.txt
numpy>=1.26.0
opencv-python>4.8.0
open3d>=0.17.1
scipy>=1.12.0
PyYAML>=6.0.0
scikit-learn>=1.3.0
gtsam-develop


